update test structures
